Struct axdiskfs::dir::Dir

pub struct Dir {
    pub entries: Vec<DirEntry>,
    /* private fields */
}

Basic Dir Struct for DirNode

Fields
entries: Vec<DirEntry>

A vector of DirEntry, each entry is a DirEntry

Implementations

impl Dir

pub fn new_root(entries: &Vec<DirEntry>) -> Self

Use available entries to create a new root Dir Struct.

pub fn new(first_cluster: u32, parent_first_cluster: u32) -> Self

Create a new Dir Struct. Unlike root, the ordinary dir needs to initialize the first and second entry.

pub fn get_entry_by_index(&self, index: usize) -> Option<DirEntry>

Return the entry by index.

pub fn get_self_first_cluster(&self) -> u32

Return the first cluster used by itself.

pub fn find_next_free_entry(&self) -> Option<usize>

Find the next free entry in entries

pub fn get_entry_by_name(&self, name: &str) -> Option<usize>

Return the entry by name.

pub fn set_entries(&mut self, entries: Vec<DirEntry>)

Set the entries of Dir.

pub fn update_file_size( &mut self, file_name: &str, file_size: u32 ) -> Result<(), DevError>

Find entry by name, if not found, return Err, else update file_size and return Ok.

pub fn update_entries_from_disk(&mut self) -> Result<(), DevError>

Read all entries from disk and update entries.

pub fn write_entries_to_disk(&mut self) -> Result<(), DevError>

Write all entries of current memory to disk.

pub fn add_entry(&mut self, entry: DirEntry) -> Result<(), DevError>

Add a entry to entries, find a entry that name[0] == 0xE5 or name[0] == 0x00, if can’t find then create a new one after the last entry

pub fn delete_entry(&mut self, name: &str) -> Result<(), DevError>

Delete the entry by name, set name[0] = 0xE5, then free cluster.

pub fn is_entry_dir(&self, name: &str) -> Result<bool, DevError>

Find the entry by name, return true if entry is dir, else return false.

pub fn update_entry( &mut self, index: u32, entry: DirEntry ) -> Result<(), DevError>

Update the entire entry by index.

pub fn update_entry_name( &mut self, original_name: &str, target_name: &str ) -> Result<(), DevError>

Update entry’s name, if can’t find the entry, return Err