Struct axdiskfs::diskfs::CCFileSystem

pub struct CCFileSystem { /* private fields */ }

A abstraction Struct of a FAT32 filesystem.

Implementations

impl CCFileSystem

pub fn new(sector_manager: Option<SectorManager>) -> Self

Create a new instance of CCFileSystem.

pub fn bytes_per_cluster(&self) -> u32

Return the bytes per cluster.

pub fn init(&self) -> Result<(), DevError>

Init the filesystem by reading boot sector and sector manager, then initialize the root dir and the fat table.

pub fn get_next_free_cluster(&self) -> u32

Return the next free cluster.

pub fn root_dir_node(&self) -> Option<Arc<DirNode>>

Returns the reference of root directory node.

pub fn get_fat_entry(&self, cluster_id: u32) -> Result<u32, DevError>

Return the fat entry via cluster_id.

pub fn read_cluster(&self, cluster_id: u32) -> Result<Vec<u8>, DevError>

Read a cluster from disk via cluster_id.

pub fn is_end(&self, index: u32) -> bool

Return true if the cluster is the end of the chain.

pub fn is_bad_cluster(&self, index: u32) -> bool

Return true if the cluster is bad.

pub fn write_cluster( &self, cluster_id: u32, data: &[u8] ) -> Result<(), DevError>

Write a cluster to disk via cluster_id.

pub fn flush_fs_info_sector(&self) -> Result<(), DevError>

Write back fs_info_sector of memory to disk.

pub fn allocate_cluster_at_middle( &self, curr_cluster_id: u32, next_cluster_id: u32 ) -> Option<u32>

Allocate a cluster between curr_cluster_id and next_cluster_id.

pub fn allocate_cluster_at_end(&self, curr_cluster_id: u32) -> Option<u32>

Allocate a cluster at the end of the chain.

pub fn allocate_cluster_at_start(&self) -> Option<u32>

Allocate a cluster at the start of the chain.

pub fn link_to_end(&self, curr_cluster_id: u32) -> Result<(), DevError>

Link the cluster to the end of the chain.

pub fn free_cluster(&self, cluster_id: u32) -> Result<(), DevError>

Free the cluster, then update the free_cluster_count and next_free_cluster.